+++ reqlient/async_/__init__.py
@@ -1,5 +1,6 @@
 """Asynchronous REST client modules."""
 
+from .bulkhead import AsyncBulkhead, AsyncBulkheadRegistry
 from .circuit_breakers import AsyncCircuitBreaker, AsyncCircuitBreakerRegistry
 from .interceptors import AsyncInterceptor, TraceContextInterceptor
 from .rest_client import AsyncRestClient
@@ -8,6 +9,8 @@ __all__ = [
     "AsyncRestClient",
     "AsyncCircuitBreaker",
     "AsyncCircuitBreakerRegistry",
+    "AsyncBulkhead",
+    "AsyncBulkheadRegistry",
     "AsyncInterceptor",
     "TraceContextInterceptor",
 ]