+++ reqlient/sync/__init__.py
@@ -1,5 +1,6 @@
 """Synchronous REST client modules."""
 
+from .bulkhead import Bulkhead, BulkheadRegistry
 from .circuit_breakers import CircuitBreakerRegistry
 from .interceptors import Interceptor
 from .rest_client import RestClient
@@ -7,5 +8,7 @@ from .rest_client import RestClient
 __all__ = [
     "RestClient",
     "CircuitBreakerRegistry",
+    "Bulkhead",
+    "BulkheadRegistry",
     "Interceptor",
 ]